# packages/shared/schema/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_voice_sample_v1_voice_profiles_upload_sample_post": {
        "properties": {
          "sample": {
            "description": "음성 샘플 파일",
            "format": "binary",
            "title": "Sample",
            "type": "string"
          }
        },
        "required": [
          "sample"
        ],
        "title": "Body_upload_voice_sample_v1_voice_profiles_upload_sample_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProfileCreateRequest": {
        "properties": {
          "age_band": {
            "default": "5-7",
            "pattern": "^(3-5|5-7|7-9|adult)$",
            "title": "Age Band",
            "type": "string"
          },
          "avatar_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "birth_month": {
            "anyOf": [
              {
                "maximum": 12.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Month"
          },
          "birth_year": {
            "anyOf": [
              {
                "maximum": 2100.0,
                "minimum": 2005.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Year"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "name": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "preferred_theme": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Theme"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProfileCreateRequest",
        "type": "object"
      },
      "ProfileUpdateRequest": {
        "properties": {
          "age_band": {
            "anyOf": [
              {
                "pattern": "^(3-5|5-7|7-9|adult)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Band"
          },
          "avatar_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "birth_month": {
            "anyOf": [
              {
                "maximum": 12.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Month"
          },
          "birth_year": {
            "anyOf": [
              {
                "maximum": 2100.0,
                "minimum": 2005.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Year"
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "preferred_theme": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Theme"
          }
        },
        "title": "ProfileUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VoiceProfileCreateRequest": {
        "properties": {
          "consented": {
            "default": false,
            "title": "Consented",
            "type": "boolean"
          },
          "label": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "provider_voice_id": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Voice Id"
          },
          "relationship": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship"
          },
          "sample_audio_url": {
            "maxLength": 500,
            "minLength": 8,
            "title": "Sample Audio Url",
            "type": "string"
          }
        },
        "required": [
          "label",
          "sample_audio_url"
        ],
        "title": "VoiceProfileCreateRequest",
        "type": "object"
      },
      "VoiceProfileUpdateRequest": {
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "consented": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consented"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "provider_voice_id": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Voice Id"
          },
          "relationship": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship"
          },
          "sample_audio_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Audio Url"
          }
        },
        "title": "VoiceProfileUpdateRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@aistorybook.com",
      "name": "AI Story Book Team"
    },
    "description": "\n# AI Story Book API\n\nAI 기반 맞춤형 동화책 생성 API입니다.\n\n## 주요 기능\n\n* **책 생성**: 주제, 연령대, 스타일을 입력하면 AI가 동화책을 생성합니다\n* **캐릭터 관리**: 사진에서 캐릭터를 추출하거나 직접 생성할 수 있습니다\n* **시리즈 생성**: 같은 캐릭터로 연속된 이야기를 만들 수 있습니다\n* **PDF/오디오 내보내기**: 완성된 책을 PDF나 오디오로 내보낼 수 있습니다\n\n## 인증\n\n모든 API는 `X-User-Key` 헤더가 필요합니다.\n중복 요청 방지를 위해 `X-Idempotency-Key` 헤더 사용을 권장합니다.\n\n## Rate Limiting\n\n- 기본: 분당 10회 요청 제한\n- 초과 시 429 Too Many Requests 응답\n\n## 크레딧 시스템\n\n책 1권 생성에 크레딧 1개가 소모됩니다.\n    ",
    "license": {
      "name": "MIT"
    },
    "title": "AI Story Book API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/profiles": {
      "get": {
        "operationId": "list_profiles_v1_profiles_get",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Profiles",
        "tags": [
          "Profiles"
        ]
      },
      "post": {
        "operationId": "create_profile_v1_profiles_post",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Profile",
        "tags": [
          "Profiles"
        ]
      }
    },
    "/v1/profiles/{profile_id}": {
      "delete": {
        "operationId": "delete_profile_v1_profiles__profile_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Profile",
        "tags": [
          "Profiles"
        ]
      },
      "patch": {
        "operationId": "update_profile_v1_profiles__profile_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile",
        "tags": [
          "Profiles"
        ]
      }
    },
...
    "/v1/voice-profiles": {
      "get": {
        "operationId": "list_voice_profiles_v1_voice_profiles_get",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Voice Profiles",
        "tags": [
          "VoiceProfiles"
        ]
      },
      "post": {
        "operationId": "create_voice_profile_v1_voice_profiles_post",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceProfileCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Voice Profile",
        "tags": [
          "VoiceProfiles"
        ]
      }
    },
    "/v1/voice-profiles/upload-sample": {
      "post": {
        "operationId": "upload_voice_sample_v1_voice_profiles_upload_sample_post",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_voice_sample_v1_voice_profiles_upload_sample_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Voice Sample",
        "tags": [
          "VoiceProfiles"
        ]
      }
    },
    "/v1/voice-profiles/{profile_id}": {
      "delete": {
        "operationId": "delete_voice_profile_v1_voice_profiles__profile_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Voice Profile",
        "tags": [
          "VoiceProfiles"
        ]
      },
      "patch": {
        "operationId": "update_voice_profile_v1_voice_profiles__profile_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Voice Profile",
        "tags": [
          "VoiceProfiles"
        ]
      }
    },
    "/v1/voice-profiles/{profile_id}/revoke-consent": {
      "post": {
        "operationId": "revoke_voice_profile_consent_v1_voice_profiles__profile_id__revoke_consent_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Voice Profile Consent",
        "tags": [
          "VoiceProfiles"
        ]
      }
    }
  },
...
}